integration: Add test for io command
