+++ integration/testdata/helloworld/uinit/helloworld.go
@@ -1,3 +1,7 @@
+// Copyright 2018 the u-root Authors. All rights reserved
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 package main
 
 import (
