+++ integration/uinit_test.go
@@ -1,0 +1,31 @@
+// Copyright 2018 the u-root Authors. All rights reserved
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package integration
+
+import (
+	"testing"
+)
+
+// TestHelloWorld runs an init which prints the string "HELLO WORLD" and exits.
+func TestHelloWorld(t *testing.T) {
+	// Create the CPIO and start QEMU.
+	tmpDir, q := testWithQEMU(t, "github.com/u-root/u-root/integration/testdata/helloworld/uinit")
+	defer cleanup(t, tmpDir, q)
+
+	if err := q.Expect("HELLO WORLD"); err != nil {
+		t.Fatal(err)
+	}
+}
+
+// TestHelloWorldNegative runs an init which does not print the string "HELLO WORLD".
+func TestHelloWorldNegative(t *testing.T) {
+	// Create the CPIO and start QEMU.
+	tmpDir, q := testWithQEMU(t, "github.com/u-root/u-root/integration/testdata/helloworld/uinit")
+	defer cleanup(t, tmpDir, q)
+
+	if err := q.Expect("GOODBYE WORLD"); err == nil {
+		t.Fatal(err)
+	}
+}
